+++ opentelemetry-api/src/opentelemetry/trace/span.py
@@ -86,6 +86,7 @@ class Span(abc.ABC):
         exception: Exception,
         attributes: types.Attributes = None,
         timestamp: typing.Optional[int] = None,
+        escaped: bool = False,
     ) -> None:
         """Records an exception as a span event."""
 
@@ -189,6 +190,17 @@ class SpanContext(
             (trace_id, span_id, is_remote, trace_flags, trace_state, is_valid),
         )
 
+    def __getnewargs__(
+        self,
+    ) -> typing.Tuple[int, int, bool, "TraceFlags", "TraceState"]:
+        return (
+            self.trace_id,
+            self.span_id,
+            self.is_remote,
+            self.trace_flags,
+            self.trace_state,
+        )
+
     @property
     def trace_id(self) -> int:
         return self[0]  # pylint: disable=unsubscriptable-object
@@ -275,6 +287,7 @@ class DefaultSpan(Span):
         exception: Exception,
         attributes: types.Attributes = None,
         timestamp: typing.Optional[int] = None,
+        escaped: bool = False,
     ) -> None:
         pass